Seperate base from derived classes for pluginlibs
