+++ core/src/stages/grasp_provider_base.cpp
@@ -1,0 +1,76 @@
+//
+// Created by jafar_abdi on 3/27/21.
+//
+
+#include <pluginlib/class_list_macros.h>
+#include <eigen_conversions/eigen_msg.h>
+#include <Eigen/Geometry>
+#include <moveit/planning_scene/planning_scene.h>
+#include <rviz_marker_tools/marker_creation.h>
+#include <moveit/task_constructor/marker_tools.h>
+#include <moveit/task_constructor/storage.h>
+#include <moveit/task_constructor/stages/grasp_provider.h>
+#include <moveit/task_constructor/stages/grasp_provider_base.h>
+
+namespace moveit {
+namespace task_constructor {
+namespace stages {
+GraspProviderBase::GraspProviderBase(const std::string& name) : GeneratePose(name) {
+	auto& p = properties();
+	p.declare<std::string>("eef", "name of end-effector");
+	p.declare<std::string>("object");
+
+	p.declare<boost::any>("pregrasp", "pregrasp posture");
+	p.declare<boost::any>("grasp", "grasp posture");
+}
+void GraspProviderBase::init(const std::shared_ptr<const moveit::core::RobotModel>& robot_model) {
+	InitStageException errors;
+	try {
+		GeneratePose::init(robot_model);
+	} catch (InitStageException& e) {
+		errors.append(e);
+	}
+
+	const auto& props = properties();
+
+	// check availability of object
+	props.get<std::string>("object");
+	// check availability of eef
+	const std::string& eef = props.get<std::string>("eef");
+	if (!robot_model->hasEndEffector(eef))
+		errors.push_back(*this, "unknown end effector: " + eef);
+	else {
+		// check availability of eef pose
+		const moveit::core::JointModelGroup* jmg = robot_model->getEndEffector(eef);
+		const std::string& name = props.get<std::string>("pregrasp");
+		std::map<std::string, double> m;
+		if (!jmg->getVariableDefaultPositions(name, m))
+			errors.push_back(*this, "unknown end effector pose: " + name);
+	}
+
+	if (errors)
+		throw errors;
+}
+void GraspProviderBase::onNewSolution(const SolutionBase& s) {
+	std::shared_ptr<const planning_scene::PlanningScene> scene = s.end()->scene();
+
+	const auto& props = properties();
+	const std::string& object = props.get<std::string>("object");
+	if (!scene->knowsFrameTransform(object)) {
+		const std::string msg = "object '" + object + "' not in scene";
+		if (storeFailures()) {
+			InterfaceState state(scene);
+			SubTrajectory solution;
+			solution.markAsFailure();
+			solution.setComment(msg);
+			spawn(std::move(state), std::move(solution));
+		} else
+			ROS_WARN_STREAM_NAMED("GraspProviderBase", msg);
+		return;
+	}
+
+	upstream_solutions_.push(&s);
+}
+}  // namespace stages
+}  // namespace task_constructor
+}  // namespace moveit
